Damysos

Documentation for Damysos.

Damysos.Ensemble

Damysos.GaussianAPulse

Damysos.GaussianEPulse

Damysos.Simulation

Damysos.UnitScaling

Damysos.run_simulation!

Damysos.run_simulation!

Damysos.run_simulation1d!

Damysos.run_simulation2d!

Damysos.Ensemble — Type

Ensemble{T}(simlist, id, datapath, plotpath)

A struct representing an ensemble of simulations.

Damysos.GaussianAPulse — Type

GaussianAPulse{T<:Real}

Represents spacially homogeneous, linearly polarized pulse with Gaussian envelope.

Mathematical form

The form of the vector potential is given by

\[\vec{A}(t) = \vec{A}_0 \cos(\omega t) e^{-t^2 / \sigma^2}\]

where $\vec{A}_0=A_0(\cos\varphi\,\vec{e}_x + \sin\varphi\,\vec{e}_y$).

Damysos.GaussianEPulse — Type

GaussianEPulse{T<:Real}

Represents spacially homogeneous, linearly polarized pulse with Gaussian envelope.

Mathematical form

The form of the electric field is given by

\[\vec{E}(t) = \vec{E}_0 \sin(\omega t+\phi) e^{-t^2 / \sigma^2}\]

where $\vec{E}_0=E_0(\cos\varphi\,\vec{e}_x + \sin\varphi\,\vec{e}_y$).

Damysos.Simulation — Type

Simulation{T}(hamiltonian, drivingfield, numericalparams, observables, unitscaling, dimensions, id, datapath, plotpath)

A struct representing a simulation with various components.

Damysos.UnitScaling — Type

UnitScaling(timescale,lengthscale)

Represents a physical length- and time-scale used for non-dimensionalization of a system.

Examples

julia> using Unitful; us = UnitScaling(u"1.0s",u"1.0m")
UnitScaling
    timescale: 1.0e15 fs
    lengthscale: 1.0e9 nm

Further information

See here

Damysos.run_simulation! — Method

run_simulation!(ens::Ensemble;
    savedata=true,
    saveplots=true,
    ensembleparallel=false,
    threaded=false,
    maxparallel_ky=64,
    kxbatch_basesize=512,
    makecombined_plots=true,
    kwargs...)

Run simulations for an ensemble of sim objects.

Arguments

ens::Ensemble: See Ensemble

maxparallel_ky: The maximum amount of different ky-lines computed in parallel. Good values are typically ~ 2nworkers. Large numbers mean high memory footprint.

kxbatch_basesize : Number of kx modes per ky-line processed in one solve call. Large numbers mean high memory footprint.

kwargs...: Additional keyword arguments are passed to the solve() function of DifferentialEquations.jl

Returns

An array of observables obtained from the simulations.

See also

run_simulation2d!, run_simulation!

Damysos.run_simulation! — Method

run_simulation!(sim::Simulation;
    savedata=true,
    saveplots=true,
    threaded=false,
    maxparallel_ky=64,
    kxbatch_basesize=512,
    kwargs...)

Run a simulation.

Arguments

sim::Simulation: See Simulation

maxparallel_ky: The maximum amount of different ky-lines computed in parallel. Good values are typically ~ 2nworkers. Large numbers mean high memory footprint.

kxbatch_basesize : Number of kx modes per ky-line processed in one solve call. Large numbers mean high memory footprint.

kwargs...: Additional keyword arguments are passed to the solve() function of DifferentialEquations.jl

Returns

The observables obtained from the simulation.

See also

run_simulation1d!, run_simulation2d!

Damysos.run_simulation1d! — Method

run_simulation1d!(sim::Simulation{T};kwargs...)

Run a 1D simulation for a given Simulation sim.

This function should not be called by a user, since it has no convenience features such as saving and plotting. Use run_simulation! instead

Returns

The combined observables obtained from the simulation.

See also

run_simulation!, run_simulation2d!

Damysos.run_simulation2d! — Method

run_simulation2d!(sim::Simulation{T};kwargs...)

Run a 2D simulation for a given Simulation sim.

This function should not be called by a user, since it has no convenience features such as saving and plotting. Use run_simulation! instead

Returns

The combined observables obtained from the simulation.

See also

run_simulation!, run_simulation1d!