Two-band Hamiltonians

Here we outline the implementation of calculating matrixelements of two-band Hamiltonians.

Idea

Consider Hamiltonians defined in k-space which are of the general form

\[\hat{H} = \vec{h}(\vec{k})\cdot\vec{\sigma}\]

Since this is a simple 2x2 matrix, one can express all desired operator matrix elements in terms of $\vec{h}(\vec{k})$ and its derivatives.

Below we list the functions implemented so far.

Eigenstate basis

At the moment all matrix elements are given in the basis of the Hamiltonian's eigenstates describing valence- and conduction band states. Some matrix elements depend on the eigenstates' gauge, so we give the explicit form of the eigenstates used here in the following:

\[\hat{H}\ket{\psi_\mu} = \epsilon_\mu\ket{\psi_\mu}\]

with

\[\epsilon_{\mu}(\vec{k})\equiv\epsilon_{\pm}(\vec{k})=\pm\|\vec{h}(\vec{k})\|\equiv\pm h(\vec{k})\]

In $\sigma_z$-standardbasis ($\mu\in\{\uparrow,\downarrow\}$) the eigenstates read

\[\braket{\mu|\psi_+} = \frac{1}{\sqrt{h(h+h_z)}}\begin{pmatrix} h + h_z\\h_x + i\,h _y \end{pmatrix}\]

\[\braket{\mu|\psi_-} = \frac{1}{\sqrt{h(h+h_z)}}\begin{pmatrix} -h_x + i\, h_y\\h + h_z \end{pmatrix}\]

Velocity operator matrix elements

The velocity operator is given by $\hat{v}^\mu=\frac{\partial\hat{H}}{\partial k_\mu}$ where $\mu\in\{x,y,z\}$. We denote its matrix elements via

\[v_{mn}^\mu = \braket{\psi_m|\hat{v}^\mu|\psi_n}\]

where we use the indices $m,n\in\{+,-\}=\{c,v\}$ for positive (negative) and conduction (valence) band interchangeably. The matrix elements can be simplified to

\[\begin{aligned} v_{cc}^\mu(\vec{k}) &= \frac{\partial\epsilon}{\partial k_\mu}\\ v_{cv}^\mu(\vec{k}) &= \frac{\partial\vec{h}}{\partial k_\mu}\cdot\vec{\sigma}_{cv}(\vec{k})\\ v_{vc}^\mu(\vec{k}) &=\left[v_{cv}^\mu(\vec{k})\right]^*\\ v_{vv}^\mu(\vec{k}) &= -\frac{\partial\epsilon(\vec{k})}{\partial k_\mu}\\ \end{aligned}\]

using the eigenstates and the matrix elements $\vec{\sigma}_{mn}=\braket{\psi_m|\vec{\sigma}|\psi_n}$ of the Pauli matrices.

Dipole operator matrix elements

The dipole operator for a charge $q$ is given by $\hat{d}^\mu=q\,\hat{r}^\mu$ where $\mu\in\{x,y,z\}$. We denote its matrix elements via

\[d_{mn}^\mu = q\braket{\psi_m|\hat{r}^\mu|\psi_n}\]

where we use the indices $m,n\in\{+,-\}=\{c,v\}$ for positive (negative) and conduction (valence) band interchangeably. The matrix elements can be simplified to

\[\begin{aligned} d_{cc}^\mu(\vec{k}) &= \frac{1}{2\epsilon(\vec{k})(\epsilon(\vec{k}) + h_z(\vec{k}))} \left( \frac{\partial\vec{h}}{\partial k_\mu} \right)^T\Omega\,\vec{h}(\vec{k})\\ d_{cv}^\mu(\vec{k}) &= \frac{-i}{\Delta\epsilon(\vec{k})} \frac{\partial\vec{h}(\vec{k})}{\partial k_\mu}\cdot\vec{\sigma}_{cv}(\vec{k})\\ d_{vc}^\mu(\vec{k}) &=\left[v_{cv}^\mu(\vec{k})\right]^*\\ d_{vv}^\mu(\vec{k}) &= -d_{cc}^\mu(\vec{k}) \end{aligned}\]

with the matrix

\[\Omega = \begin{pmatrix} 0 & -1 & 0\\ 1 & 0 & 0\\ 0 & 0 & 0 \end{pmatrix}.\]

Implementing functions

Damysos.vx_cc — Function

vx_cc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vx|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_cc(h,1.0,-1.0)
0.5773502691896258

Damysos.vx_cv — Function

vx_cv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vx|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_cv(h,1.0,-1.0)
0.7886751345948129 - 0.21132486540518708im

Damysos.vx_vc — Function

vx_vc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vx|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_vc(h,1.0,-1.0)
0.7886751345948129 + 0.21132486540518708im

Damysos.vx_vv — Function

vx_vv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vx|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_vv(h,1.0,-1.0)
-0.5773502691896258

Damysos.vy_cc — Function

vy_cc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vy|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_cc(h,1.0,-1.0)
-0.5773502691896258

Damysos.vy_cv — Function

vy_cv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vy|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_cv(h,1.0,-1.0)
0.21132486540518708 - 0.7886751345948129im

Damysos.vy_vc — Function

vy_vc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vy|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_vc(h,1.0,-1.0)
0.21132486540518708 + 0.7886751345948129im

Damysos.vy_vv — Function

vy_vv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vy|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_vv(h,1.0,-1.0)
0.5773502691896258

Damysos.dx_cc — Function

dx_cc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|x|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_cc(h,1.0,-1.0)
-0.10566243270259357

Damysos.dx_cv — Function

dx_cv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|x|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_cv(h,1.0,-1.0)
-0.0610042339640731 - 0.22767090063073978im

Damysos.dx_vc — Function

dx_vc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|x|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_vc(h,1.0,-1.0)
-0.0610042339640731 + 0.22767090063073978im

Damysos.dx_vv — Function

dx_vv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|x|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_vv(h,1.0,-1.0)
0.10566243270259357

Damysos.dy_cc — Function

dy_cc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|y|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_cc(h,1.0,-1.0)
-0.10566243270259357

Damysos.dy_cv — Function

dy_cv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|y|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_cv(h,1.0,-1.0)
-0.22767090063073978 - 0.0610042339640731im

Damysos.dy_vc — Function

dy_vc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|y|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_vc(h,1.0,-1.0)
-0.22767090063073978 + 0.0610042339640731im

Damysos.dy_vv — Function

dy_vv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|y|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_vv(h,1.0,-1.0)
0.10566243270259357