Reference

Index

Damysos.m_e

Damysos.q_e

Damysos.ħ

Damysos.BilayerToy

Damysos.CartesianKGrid1d

Damysos.CartesianKGrid2d

Damysos.ConvergenceTest

Damysos.ConvergenceTestResult

Damysos.DensityMatrix

Damysos.DensityMatrixSnapshots

Damysos.ExtendKymaxTest

Damysos.GappedDirac

Damysos.GaussianAPulse

Damysos.GaussianEPulse

Damysos.GeneralTwoBand

Damysos.KGrid0d

Damysos.KGridEmpty

Damysos.LinearCUDA

Damysos.LinearChunked

Damysos.LinearTest

Damysos.NGrid

Damysos.Occupation

Damysos.PowerLawTest

Damysos.QuadraticToy

Damysos.Simulation

Damysos.SingleMode

Damysos.SymmetricTimeGrid

Damysos.TwoBandDephasingLiouvillian

Damysos.UnitScaling

Damysos.Vectorpotential

Damysos.VectorpotentialFreeUnits

Damysos.VectorpotentialUnits

Damysos.Velocity

Damysos.VelocityX

Damysos.buildrhs_cc_cv_x_expression

Damysos.buildrhs_x_expression_svec

Damysos.define_functions

Damysos.dx_cc

Damysos.dx_cv

Damysos.dx_vc

Damysos.dx_vv

Damysos.dy_cc

Damysos.dy_cv

Damysos.dy_vc

Damysos.dy_vv

Damysos.estimate_atol

Damysos.hmat

Damysos.hvec

Damysos.linear_fit

Damysos.run!

Damysos.run!

Damysos.successful_retcode

Damysos.successful_retcode

Damysos.terminated_retcode

Damysos.terminated_retcode

Damysos.vx_cc

Damysos.vx_cv

Damysos.vx_vc

Damysos.vx_vv

Damysos.vy_cc

Damysos.vy_cv

Damysos.vy_vc

Damysos.vy_vv

Damysos.Δϵ

Damysos.ϵ

Docstrings

Damysos.Vectorpotential — Type

Damysos.Vectorpotential{T, U}

A supertype for quantities and levels of dimension 𝐋 * 𝐌 * 𝐓 ^ -2 * 𝐈 ^ -1 with a value of type T and units U.

See also: Unitful.Quantity, Unitful.Level.

Damysos.m_e — Constant

Rest mass of an electron as Quantity (Unitful.jl package) equal to 9.1093837139(28)e-31kg

Damysos.q_e — Constant

Elementary charge as Quantity (Unitful.jl package) equal to 1.602176634e-19C

Damysos.ħ — Constant

Reduced Planck constant as Quantity (Unitful.jl package) equal to 6.582119569...e-16 eV⋅s

Damysos.BilayerToy — Type

BilayerToy{T<:Real} <: GeneralTwoBand{T}

A toy model of touching quadratic bands with a gap turning them quartic near the band edge.

The Hamiltonian reads

\[\hat{H} = \frac{\zeta}{2}[(k_y^2-k_x^2)\sigma_x - 2 k_x k_y \sigma_y] + \frac{\Delta}{2}\sigma_z\]

such that $\vec{h}=[\zeta/2 (k_y^2-k_x^2), -\zeta k_x k_y, \Delta/2]$. The dimensionful form (SI) would be

\[\hat{H}_{SI} = \frac{\hbar^2}{2m^*}[(k_y^2-k_x^2)\sigma_x - 2 k_x k_y \sigma_y]+\frac{E_{gap}}{2}\sigma_z\]

Examples

julia> h = BilayerToy(0.2,1.0)
BilayerToy:
  Δ: 0.2
  ζ: 1.0
 

See also

GeneralTwoBand GappedDirac

Damysos.CartesianKGrid1d — Type

CartesianKGrid1d{T}(dkx, kxmax[, ky]) <: CartesianKGrid{T}

One-dimensional equidistant samples in k-space in kx direction at ky.

See also

Simulation, SymmetricTimeGrid, KGrid0d, CartesianKGrid2d

Damysos.CartesianKGrid2d — Type

CartesianKGrid2d{T}(dkx, kxmax, dky, kymax) <: CartesianKGrid{T}

Two-dimensional equidistant cartesian grid in reciprocal (k-)space.

See also

Simulation, SymmetricTimeGrid, CartesianKGrid1d, KGrid0d

Damysos.ConvergenceTest — Type

ConvergenceTest(start, solver::DamysosSolver = LinearChunked(); kwargs...)

A convergence test based on a Simulation and a ConvergenceTestMethod.

Arguments

start: the starting point can be a Simulation object or are path to a previous test

solver::DamysosSolver = LinearChunked(): the solver used for simulations,

Keyword arguments

method::ConvergenceTestMethod: specifies the convergence parmeter & iteration method

resume::Bool: if true re-use the completedsims, otherwise start from scratch

atolgoal::Real: desired absolute tolerance

rtolgoal::Real: desired relative tolerance

maxtime::Union{Real,Unitful.Time}: test guaranteed to stop after maxtime

maxiterations::Integer: test stops after maxiterations Simulations were performed

path::String: path to save data of convergence test

altpath: path to try inf start.datapath throws an error

See also

LinearTest, PowerLawTest, Simulation

Damysos.ConvergenceTestResult — Type

Result of a ConvergenceTest

Damysos.DensityMatrix — Type

DensityMatrix{T<:Real,U}

Holds the density matrix elements on a k grid.

Fields

kgrid::KGrid{T}: k grid on which the density matrix is defined

density_matrix::Vector{U}: density matrices at each k-point

See also

DensityMatrixSnapshots

Damysos.DensityMatrixSnapshots — Type

DensityMatrixSnapshots{T<:Real,U} <: Observable{T}

Holds snapshots in time of the density matrix on the k grid.

Fields

tsamples::Vector{T}: time samples at which the density matrix is stored

density_matrices::Vector{DensityMatrix{T,U}}: vector of density matrices of type U

See also

Occupation, Velocity, DensityMatrix

Damysos.ExtendKymaxTest — Type

ExtendKymaxTest(extendmethod::ConvergenceTestMethod)

Specialized for extending the integration region in ky-direction avoided re-calculation.

See also

ConvergenceTest, LinearTest PowerLawTest

Damysos.GappedDirac — Type

GappedDirac{T<:Real} <: GeneralTwoBand{T}

Massive Dirac Hamiltonian (two-band model).

The Hamiltonian reads

\[\hat{H} = k_x\sigma_x + k_y\sigma_y + m\sigma_z\]

such that $\vec{h}=[k_x,k_y,m]$.

Examples

julia> h = GappedDirac(1.0)
GappedDirac:
  m: 1.0
  vF: 1.0

See also

GeneralTwoBand QuadraticToy BilayerToy

Damysos.GaussianAPulse — Type

GaussianAPulse{T<:Real}

Represents spacially homogeneous, linearly polarized pulse with Gaussian envelope.

Mathematical form

The form of the vector potential is given by

\[\vec{A}(t) = \vec{A}_0 \cos(\omega t + \theta) e^{-t^2 / 2\sigma^2}\]

where $\vec{A}_0=A_0(\cos\varphi\,\vec{e}_x + \sin\varphi\,\vec{e}_y$).

Damysos.GaussianEPulse — Type

GaussianEPulse{T<:Real}

Represents spacially homogeneous, linearly polarized pulse with Gaussian envelope.

Mathematical form

The form of the electric field is given by

\[\vec{E}(t) = \vec{E}_0 \sin(\omega t+\phi) e^{-t^2 / \sigma^2}\]

where $\vec{E}_0=E_0(\cos\varphi\,\vec{e}_x + \sin\varphi\,\vec{e}_y$).

Damysos.GeneralTwoBand — Type

GeneralTwoBand{T} <: Hamiltonian{T}

Supertype of all 2x2 Hamiltonians with all matrixelements via dispatch.

Idea

The idea is that all Hamiltonians of the form

\[\hat{H} = \vec{h}(\vec{k})\cdot\vec{\sigma}\]

can be diagonalized analytically and hence most desired matrixelements such as velocities or dipoles can be expressed solely through

\[\vec{h}(\vec{k})=[h_x(\vec{k}),h_y(\vec{k}),h_z(\vec{k})]\]

and its derivatives with respect to $k_\mu$. Any particular Hamiltonian deriving form GeneralTwoBand{T} must then only implement $\vec{h}(\vec{k})$ and its derivatives.

See also

GappedDirac

Damysos.KGrid0d — Type

KGrid0d{T}(kx,ky) <: CartesianKGrid{T}

Zero-dimensional grid, i.e. a single point in k-space.

See also

Simulation, SymmetricTimeGrid, CartesianKGrid1d, CartesianKGrid2d

Damysos.KGridEmpty — Type

Empty k grid for initialization purposes

Damysos.LinearCUDA — Type

LinearCUDA

Represents an integration strategy for k-space via simple midpoint sum, where individual k points are computed concurrently on one or several CUDA GPU(s) via linear indexing.

Fields

kchunksize::Int64: number of k-points in one concurrently executed chunk.

algorithm::GPUODEAlgorithm: algorithm for solving differential equations

ngpus::Int: #GPUs to use, automatically chooses all available GPUs if none given

rtol::Union{Nothing, Real}: relative tolerance of solver (nothing for fixed-timestep)

atol::Union{Nothing, Real}: absolute tolerance of solver (nothing for fixed-timestep)

See also

LinearChunked, SingleMode

Damysos.LinearChunked — Type

LinearChunked

Represents an integration strategy for k-space via simple midpoint sum.

Fields

kchunksize::T: number of k-points in one chunk. Every task/worker gets one chunk.

algorithm::SciMLBase.BasicEnsembleAlgorithm: algorithm for the EnsembleProblem.

odesolver::SciMLBase.AbstractODEAlgorithm: ODE algorithm

Examples

julia> solver = LinearChunked(256,EnsembleThreads())
LinearChunked:
  - kchunksize: 256
  - algorithm: EnsembleThreads()
  - odesolver: Vern7{typeof(OrdinaryDiffEqCore.trivial_limiter!), typeof(OrdinaryDiffEqCore.trivial_limiter!), Static.False}(OrdinaryDiffEqCore.trivial_limiter!, OrdinaryDiffEqCore.trivial_limiter!, static(false), true)
  

See also

LinearChunked, SingleMode

Damysos.LinearTest — Type

LinearTest{T<:Real}(parameter::Symbol,shift{T})

A convergence method where parameter is changed by adding shift each iteration.

See also

ConvergenceTest, PowerLawTest ExtendKymaxTest

Damysos.NGrid — Type

NGrid{T}(kgrid, tgrid)

Represents the discretzation of a Simulation in reciprocal (k-)space and time.

Fields

kgrid::KGrid{T}: set of points in k-space used for integration of observables.

tgrid::TimeGrid{T}: points in time-space to evaluate observables.

See also

Simulation, SymmetricTimeGrid, CartesianKGrid1d, CartesianKGrid2d KGrid0d

Damysos.Occupation — Type

Occupation{T<:Real} <: Observable{T}

Holds time series data of the occupation computed from the density matrix.

Only the conduction band occupation $\rho_{cc}(t)$ is stored since $Tr\rho(t)=1$

Fields

cbocc::Vector{T}: time-dependent conduction band occupation $\rho_{cc}(t)$

See also

Velocity

Damysos.PowerLawTest — Type

PowerLawTest{T<:Real}(parameter::Symbol,multiplier{T})

A convergence method multiplying parameter by multiplier each iteration.

See also

ConvergenceTest, LinearTest ExtendKymaxTest

Damysos.QuadraticToy — Type

QuadraticToy{T<:Real} <: GeneralTwoBand{T}

A toy model of quadratic bands with a gap.

The Hamiltonian reads

\[\hat{H} = \frac{\zeta}{2}(k_x^2\sigma_x + k_y^2\sigma_y) + \frac{\Delta}{2}\sigma_z\]

such that $\vec{h}=[\zeta/2 k_x^2,\zeta/2 k_y^2, \Delta/2]$. The dimensionful form (SI) would be

\[\hat{H}_{SI} = \frac{\hbar^2}{2m^*}(k_x^2\sigma_x+k_y^2\sigma_y)+\frac{E_{gap}}{2}\sigma_z\]

Examples

julia> h = QuadraticToy(0.2,1.0)
QuadraticToy:
  Δ: 0.2
  ζ: 1.0
 

See also

GeneralTwoBand GappedDirac

Damysos.Simulation — Type

Simulation{T}(l, df, g, obs, us, d[, id])

Represents a simulation with all physical and numerical parameters specified.

Fields

l::Liouvillian{T}: describes physical system via Liouville operator

df::DrivingField{T}: laser field driving the system

g::NGrid{T}: time & reciprocal (k-) space discretization

obs::Vector{Observable{T}}: physical observables to be computed

us::UnitScaling{T}: time- and lengthscale linking dimensionless units to SI units

id::String: identifier of the Simulation

dimensions::UInt8: system can be 0d (single mode),1d or 2d

See also

NGrid, TwoBandDephasingLiouvillian, UnitScaling, Velocity, Occupation, GaussianAPulse

Damysos.SingleMode — Type

SingleMode

Represents the solver for a single point in k-space.

See also

LinearChunked, LinearCUDA

Damysos.SymmetricTimeGrid — Type

SymmetricTimeGrid{T}(dt, t0)

Time discretzation of a Simulation spanning (-t0,t0) in steps of dt.

Fields

dt::T: timestep in internal dimensionless units (see UnitScaling).

t0::T: .

See also

Simulation, SymmetricTimeGrid, CartesianKGrid1d, CartesianKGrid2d

Damysos.TwoBandDephasingLiouvillian — Type

TwoBandDephasingLiouvillian{T<:Real} <: Liouvillian{T}

Represents a system with a two-band Hamiltonian and $T_2$ dephasing and $T_1$ relaxation

See also

GappedDirac, GeneralTwoBand

Damysos.UnitScaling — Type

UnitScaling(timescale,lengthscale)

Represents a physical length- and time-scale used for non-dimensionalization of a system.

Examples

julia> us = UnitScaling(u"1.0s",u"1.0m")
UnitScaling:
 timescale: 1.0e15 fs
 lengthscale: 1.0e9 nm

Further information

Internally, the fields timescale & lengthscale of UnitScaling are saved in femtoseconds and nanometers, but never used for numerical calculations. They are only needed to convert to dimensionful quantities again (Unitful package used, supports SI,cgs,... units) See here for more information on non-dimensionalization.

Damysos.VectorpotentialFreeUnits — Type

Damysos.VectorpotentialFreeUnits{U}

A supertype for Unitful.FreeUnits of dimension 𝐋 * 𝐌 * 𝐓 ^ -2 * 𝐈 ^ -1. Equivalent to Unitful.FreeUnits{U, 𝐋 * 𝐌 * 𝐓 ^ -2 * 𝐈 ^ -1}.

Damysos.VectorpotentialUnits — Type

Damysos.VectorpotentialUnits{U}

A supertype for units of dimension 𝐋 * 𝐌 * 𝐓 ^ -2 * 𝐈 ^ -1. Equivalent to Unitful.Units{U, 𝐋 * 𝐌 * 𝐓 ^ -2 * 𝐈 ^ -1}.

See also: Unitful.Units.

Damysos.Velocity — Type

Velocity{T<:Real} <: Observable{T}

Holds time series data of the physical velocity computed from the density matrix.

The velocity is computed via

\[\vec{v}(t) = Tr [\rho(t) \frac{\partial H}{\partial\vec{k}}] = \rho_{cc}(t) \vec{v}_{cc}(t) + \rho_{cv}(t) \vec{v}_{vc}(t) + \rho_{vv}(t) \vec{v}_{vv}(t) + \rho_{vc}(t) \vec{v}_{cv}(t) \]

Fields

vx::Vector{T}: total velocity in x-direction.

vxintra::Vector{T}: $\rho_{cc}(t) v^{x}_{cc}(t) + \rho_{vv}(t) v^{x}_{vv}(t)$

vxinter::Vector{T}: $\rho_{cv}(t) v^{x}_{vc}(t) + \rho_{vc}(t) v^{x}_{cv}(t)$

vy::Vector{T}: total velocity in y-direction.

vyintra::Vector{T}: $\rho_{cc}(t) v^{y}_{cc}(t) + \rho_{vv}(t) v^{y}_{vv}(t)$

vyinter::Vector{T}: $\rho_{cv}(t) v^{y}_{vc}(t) + \rho_{vc}(t) v^{y}_{cv}(t)$

See also

Occupation

Damysos.VelocityX — Type

VelocityX{T<:Real} <: Observable{T}

Holds time series data of the physical velocity in x direction.

The velocity is computed via

\[\vec{v}(t) = Tr [\rho(t) \frac{\partial H}{\partial\vec{k}}] = \rho_{cc}(t) \vec{v}_{cc}(t) + \rho_{cv}(t) \vec{v}_{vc}(t) + \rho_{vv}(t) \vec{v}_{vv}(t) + \rho_{vc}(t) \vec{v}_{cv}(t) \]

Fields

vx::Vector{T}: total velocity in x-direction.

vxintra::Vector{T}: $\rho_{cc}(t) v^{x}_{cc}(t) + \rho_{vv}(t) v^{x}_{vv}(t)$

vxinter::Vector{T}: $\rho_{cv}(t) v^{x}_{vc}(t) + \rho_{vc}(t) v^{x}_{cv}(t)$

See also

Occupation Velocity

Damysos.buildrhs_cc_cv_x_expression — Method

buildrhs_cc_cv_x_expression(s::Simulation)

Build symbolic right-hand sides for the occupations (cc) and coherences (cv).

Damysos.buildrhs_x_expression_svec — Method

buildrhs_x_expression(s::Simulation)

Construct the symbolic right-hand side for the x-direction equations of motion, returns SA.

Damysos.define_functions — Function

define_functions(sim,solver)

Hardcode the functions needed to run the Simulation.

Arguments

sim::Simulation: contains physical & numerical information (see Simulation)

solver: strategy for integrating in k-space. Defaults to LinearChunked)

Returns

Vector of functions used by run!.

See also

Simulation, run!, LinearChunked

Damysos.dx_cc — Method

dx_cc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|x|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_cc(h,1.0,-1.0)
-0.10566243270259357

Damysos.dx_cv — Method

dx_cv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|x|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_cv(h,1.0,-1.0)
-0.0610042339640731 - 0.22767090063073978im

Damysos.dx_vc — Method

dx_vc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|x|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_vc(h,1.0,-1.0)
-0.0610042339640731 + 0.22767090063073978im

Damysos.dx_vv — Method

dx_vv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|x|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_vv(h,1.0,-1.0)
0.10566243270259357

Damysos.dy_cc — Method

dy_cc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|y|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_cc(h,1.0,-1.0)
-0.10566243270259357

Damysos.dy_cv — Method

dy_cv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|y|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_cv(h,1.0,-1.0)
-0.22767090063073978 - 0.0610042339640731im

Damysos.dy_vc — Method

dy_vc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|y|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_vc(h,1.0,-1.0)
-0.22767090063073978 + 0.0610042339640731im

Damysos.dy_vv — Method

dy_vv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|y|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_vv(h,1.0,-1.0)
0.10566243270259357

Damysos.estimate_atol — Method

Estimate the largest numeric scale in a Hamiltonian

Damysos.hmat — Method

hmat(h,kx,ky)

Returns matrixelements of Hamiltonian h at $\vec{k}=[k_x,k_y]$ (in diabatic basis).

Damysos.hvec — Method

hvec(h,kx,ky)

Returns $\vec{h}(\vec{k})$ defining the Hamiltonian at $\vec{k}=[k_x,k_y]$.

Damysos.linear_fit — Method

Fits a straight line through a set of points and returns an anonymous fit-function

Damysos.run! — Function

run!(sim, functions[, solver]; kwargs...)

Run a simulation.

Arguments

sim::Simulation: contains physical & numerical information (see Simulation)

functions: needed by the solver/integrator (see define_functions).

solver: strategy for integrating in k-space. Defaults to LinearChunked)

Keyword Arguments

savedata::Bool: save observables and simulation to disk after completion

saveplots::Bool: create default plots and save them to disk after completion

savepath::String: path to directory to save data & plots

showinfo::Bool: log/display simulation info before running

nan_limit::Int: maximum tolerated number of nans in observables

Returns

The observables obtained from the simulation.

See also

Simulation, define_functions, LinearChunked

Damysos.run! — Method

run!(test::ConvergenceTest; kwargs...)

Run a convergence test and return the result as ConvergenceTestResult.

Keyword Arguments

savedata::Bool: save observables and simulations to disk

filepath::String: path to .hdf5 file to save results

nan_limit::Int: maximum tolerated number of nans in observables

max_nan_retries::Int: maximum number of iterations where nan_limit nans are tolerated

Returns

A ConvergenceTestResult object.

See also

ConvergenceTest

Damysos.successful_retcode — Method

successful_retcode(x)

Returns true if x terminated with a successful return code.

Damysos.successful_retcode — Method

successful_retcode(path::String)

Loads .hdf5 file of convergence test and returns true if it was successful.

See also

ConvergenceTest

Damysos.terminated_retcode — Method

terminated_retcode(x)

Returns true if x terminated regularly.

Damysos.terminated_retcode — Method

terminated_retcode(path::String)

Loads .hdf5 file of convergence test and returns true if it terminated regularly.

See also

ConvergenceTest

Damysos.vx_cc — Method

vx_cc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vx|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_cc(h,1.0,-1.0)
0.5773502691896258

Damysos.vx_cv — Method

vx_cv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vx|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_cv(h,1.0,-1.0)
0.7886751345948129 - 0.21132486540518708im

Damysos.vx_vc — Method

vx_vc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vx|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_vc(h,1.0,-1.0)
0.7886751345948129 + 0.21132486540518708im

Damysos.vx_vv — Method

vx_vv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vx|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_vv(h,1.0,-1.0)
-0.5773502691896258

Damysos.vy_cc — Method

vy_cc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vy|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_cc(h,1.0,-1.0)
-0.5773502691896258

Damysos.vy_cv — Method

vy_cv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vy|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_cv(h,1.0,-1.0)
0.21132486540518708 - 0.7886751345948129im

Damysos.vy_vc — Method

vy_vc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vy|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_vc(h,1.0,-1.0)
0.21132486540518708 + 0.7886751345948129im

Damysos.vy_vv — Method

vy_vv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vy|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_vv(h,1.0,-1.0)
0.5773502691896258

Damysos.Δϵ — Method

Δϵ(h,kx,ky)

Returns the band energy (valence & conduction) difference at $\vec{k}=[k_x,k_y]$.

Damysos.ϵ — Method

ϵ(h,kx,ky)

Returns the eigenenergy of the positive (conduction band) state at $\vec{k}=[k_x,k_y]$.