Damysos

Documentation for Damysos.

Damysos.ConvergenceTest

Damysos.ConvergenceTestResult

Damysos.Ensemble

Damysos.GappedDirac

Damysos.GaussianAPulse

Damysos.GaussianEPulse

Damysos.GeneralTwoBand

Damysos.LinearCUDA

Damysos.LinearChunked

Damysos.LinearTest

Damysos.Occupation

Damysos.PowerLawTest

Damysos.Simulation

Damysos.SingleMode

Damysos.TwoBandDephasingLiouvillian

Damysos.UnitScaling

Damysos.Velocity

Damysos.define_functions

Damysos.dx_cc

Damysos.dx_cv

Damysos.dx_vc

Damysos.dx_vv

Damysos.dy_cc

Damysos.dy_cv

Damysos.dy_vc

Damysos.dy_vv

Damysos.hvec

Damysos.linear_fit

Damysos.resume

Damysos.run!

Damysos.successful_retcode

Damysos.successful_retcode

Damysos.terminated_retcode

Damysos.terminated_retcode

Damysos.vx_cc

Damysos.vx_cv

Damysos.vx_vc

Damysos.vx_vv

Damysos.vy_cc

Damysos.vy_cv

Damysos.vy_vc

Damysos.vy_vv

Damysos.Δϵ

Damysos.ϵ

Damysos.ConvergenceTest — Type

ConvergenceTest(start::Simulation,
	solver::DamysosSolver = LinearChunked(),
	method::ConvergenceTestMethod = PowerLawTest(:dt, 0.5),
	atolgoal::Real = 1e-12,
	rtolgoal::Real = 1e-8,
	maxtime::Union{Real, Unitful.Time} = 600,
	maxiterations::Integer = 16;
	altpath = joinpath(pwd(), start.datapath)))

A convergence test based on a Simulation and a ConvergenceTestMethod.

Arguments

start::Simulation: the Simulation to be converged

method::ConvergenceTestMethod: specifies the convergence parmeter & iteration method

atolgoal::Real: desired absolute tolerance

rtolgoal::Real: desired relative tolerance

maxtime::Union{Real,Unitful.Time}: test guaranteed to stop after maxtime

maxiterations::Integer: test stops after maxiterations Simulations were performed

altpath: path to try inf start.datapath throws an error

See also

LinearTest, PowerLawTest, Simulation

Damysos.ConvergenceTestResult — Type

Result of a ConvergenceTest

Damysos.Ensemble — Type

Ensemble{T}(simlist, id, datapath, plotpath)

A struct representing an ensemble of simulations.

Damysos.GappedDirac — Type

GappedDirac{T<:Real} <: GeneralTwoBand{T}

Holds the parameter of a dimensionless massive Dirac Hamiltonian.

The Hamiltonian reads

\[\hat{H} = k_x\sigma_x + k_y\sigma_y + m\sigma_z\]

such that $\vec{h}=[k_x,k_y,m]$.

Examples

julia> h = GappedDirac(1.0)
GappedDirac:
 m: 1.0
 vF: 1.0

See also

GeneralTwoBand

Damysos.GaussianAPulse — Type

GaussianAPulse{T<:Real}

Represents spacially homogeneous, linearly polarized pulse with Gaussian envelope.

Mathematical form

The form of the vector potential is given by

\[\vec{A}(t) = \vec{A}_0 \cos(\omega t + \theta) e^{-t^2 / \sigma^2}\]

where $\vec{A}_0=A_0(\cos\varphi\,\vec{e}_x + \sin\varphi\,\vec{e}_y$).

Damysos.GaussianEPulse — Type

GaussianEPulse{T<:Real}

Represents spacially homogeneous, linearly polarized pulse with Gaussian envelope.

Mathematical form

The form of the electric field is given by

\[\vec{E}(t) = \vec{E}_0 \sin(\omega t+\phi) e^{-t^2 / \sigma^2}\]

where $\vec{E}_0=E_0(\cos\varphi\,\vec{e}_x + \sin\varphi\,\vec{e}_y$).

Damysos.GeneralTwoBand — Type

GeneralTwoBand{T} <: Hamiltonian{T}

Supertype of all 2x2 Hamiltonians with all matrixelements via dispatch.

Idea

The idea is that all Hamiltonians of the form

\[\hat{H} = \vec{h}(\vec{k})\cdot\vec{\sigma}\]

can be diagonalized analytically and hence most desired matrixelements such as velocities or dipoles can be expressed solely through

\[\vec{h}(\vec{k})=[h_x(\vec{k}),h_y(\vec{k}),h_z(\vec{k})]\]

and its derivatives with respect to $k_\mu$. Any particular Hamiltonian deriving form GeneralTwoBand{T} must then only implement $\vec{h}(\vec{k})$ and its derivatives.

See also

GappedDirac

Damysos.LinearCUDA — Type

LinearCUDA{T}

Represents an integration strategy for k-space via simple midpoint sum, where individual k points are computed concurrently on a CUDA GPU

Fields

kchunksize::T: number of k-points in one concurrently executed chunk.

algorithm::GPUODEAlgorithm: algorithm for solving differential equations

See also

LinearChunked, SingleMode

Damysos.LinearChunked — Type

LinearChunked{T}

Represents an integration strategy for k-space via simple midpoint sum.

Fields

kchunksize::T: number of k-points in one chunk. Every task/worker gets one chunk.

algorithm::SciMLBase.BasicEnsembleAlgorithm: algorithm for the EnsembleProblem.

odesolver::SciMLBase.AbstractODEAlgorithm: ODE algorithm

Examples

julia> solver = LinearChunked(256,EnsembleThreads())
LinearChunked:
  - kchunksize: 256
  - algorithm: EnsembleThreads()
  - odesolver: Vern7(; stage_limiter! = trivial_limiter!, step_limiter! = trivial_limiter!, thread = static(false), lazy = true,)
  

See also

LinearChunked, SingleMode

Damysos.LinearTest — Type

LinearTest{T<:Real}(parameter::Symbol,shift{T})

A convergence method where parameter is changed by adding shift each iteration.

See also

ConvergenceTest, PowerLawTest

Damysos.Occupation — Type

Occupation{T<:Real} <: Observable{T}

Holds time series data of the occupation computed from the density matrix.

Only the conduction band occupation $\rho_{cc}(t)$ is stored since $Tr\rho(t)=1$

Fields

cbocc::Vector{T}: time-dependent conduction band occupation $\rho_{cc}(t)$

See also

Velocity

Damysos.PowerLawTest — Type

PowerLawTest{T<:Real}(parameter::Symbol,multiplier{T})

A convergence method multiplying parameter by multiplier each iteration.

See also

ConvergenceTest, LinearTest

Damysos.Simulation — Type

Simulation{T}(l, df, p, obs, us, d[, id, datapath, plotpath])

Represents a simulation with all physical and numerical parameters specified.

Fields

l::Liouvillian{T}: describes physical system via Liouville operator

df::DrivingField{T}: laser field driving the system

p::NumericalParameters{T}: all numerical parameters of the system

obs::Vector{Observable{T}}: physical observables to be computed

us::UnitScaling{T}: time- and lengthscale linking dimensionless units to SI units

id::String: identifier of the Simulation

datapath::String: path to save computed observables and simulation metadata

plotpath::String: path to savee automatically generated plots

dimensions::UInt8: system can be 0d (single mode),1d or 2d

See also

Ensemble, TwoBandDephasingLiouvillian, UnitScaling, Velocity, Occupation, GaussianAPulse

Damysos.SingleMode — Type

SingleMode

Represents the solver for a single point in k-space.

See also

LinearChunked, LinearCUDA

Damysos.TwoBandDephasingLiouvillian — Type

TwoBandDephasingLiouvillian{T<:Real} <: Liouvillian{T}

Represents a system with a two-band Hamiltonian and $T_2$ dephasing and $T_1$ relaxation

See also

GappedDirac, GeneralTwoBand

Damysos.UnitScaling — Type

UnitScaling(timescale,lengthscale)

Represents a physical length- and time-scale used for non-dimensionalization of a system.

Examples

julia> using Unitful; us = UnitScaling(u"1.0s",u"1.0m")
UnitScaling:
 timescale: 1.0e15 fs
 lengthscale: 1.0e9 nm

Further information

See here

Damysos.Velocity — Type

Velocity{T<:Real} <: Observable{T}

Holds time series data of the physical velocity computed from the density matrix.

The velocity is computed via

\[\vec{v}(t) = Tr [\rho(t) \frac{\partial H}{\partial\vec{k}}] = \rho_{cc}(t) \vec{v}_{cc}(t) + \rho_{cv}(t) \vec{v}_{vc}(t) + \rho_{vv}(t) \vec{v}_{vv}(t) + \rho_{vc}(t) \vec{v}_{cv}(t) \]

Fields

vx::Vector{T}: total velocity in x-direction.

vxintra::Vector{T}: $\rho_{cc}(t) v^{x}_{cc}(t) + \rho_{vv}(t) v^{x}_{vv}(t)$

vxinter::Vector{T}: $\rho_{cv}(t) v^{x}_{vc}(t) + \rho_{vc}(t) v^{x}_{cv}(t)$

vy::Vector{T}: total velocity in y-direction.

vyintra::Vector{T}: $\rho_{cc}(t) v^{y}_{cc}(t) + \rho_{vv}(t) v^{y}_{vv}(t)$

vyinter::Vector{T}: $\rho_{cv}(t) v^{y}_{vc}(t) + \rho_{vc}(t) v^{y}_{cv}(t)$

See also

Occupation

Damysos.define_functions — Function

define_functions(sim,solver)

Hardcode the functions needed to run the Simulation.

Arguments

sim::Simulation: contains physical & numerical information (see Simulation)

solver: strategy for integrating in k-space. Defaults to LinearChunked)

Returns

Vector of functions used by run!.

See also

Simulation, run!, LinearChunked

Damysos.dx_cc — Method

dx_cc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|x|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_cc(h,1.0,-1.0)
-0.10566243270259357

Damysos.dx_cv — Method

dx_cv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|x|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_cv(h,1.0,-1.0)
-0.0610042339640731 - 0.22767090063073978im

Damysos.dx_vc — Method

dx_vc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|x|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_vc(h,1.0,-1.0)
-0.0610042339640731 + 0.22767090063073978im

Damysos.dx_vv — Method

dx_vv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|x|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_vv(h,1.0,-1.0)
0.10566243270259357

Damysos.dy_cc — Method

dy_cc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|y|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_cc(h,1.0,-1.0)
-0.10566243270259357

Damysos.dy_cv — Method

dy_cv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|y|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_cv(h,1.0,-1.0)
-0.22767090063073978 - 0.0610042339640731im

Damysos.dy_vc — Method

dy_vc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|y|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_vc(h,1.0,-1.0)
-0.22767090063073978 + 0.0610042339640731im

Damysos.dy_vv — Method

dy_vv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|y|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_vv(h,1.0,-1.0)
0.10566243270259357

Damysos.hvec — Method

hvec(h,kx,ky)

Returns $\vec{h}(\vec{k})$ defining the Hamiltonian at $\vec{k}=[k_x,k_y]$.

Damysos.linear_fit — Method

Fits a straight line through a set of points and returns an anonymous fit-function

Damysos.resume — Method

resume(filepath_hdf5::String,solver::DamysosSolver)

Returns true if x terminated with a successful return code.

Damysos.run! — Function

run!(sim, functions[, solver]; kwargs...)

Run a simulation.

Arguments

sim::Simulation: contains physical & numerical information (see Simulation)

functions: needed by the solver/integrator (see define_functions).

solver: strategy for integrating in k-space. Defaults to LinearChunked)

Keyword Arguments

savedata::Bool: save observables and simulation to disk after completion

plotdata::Bool: create default plots and save them to disk after completion

Returns

The observables obtained from the simulation.

See also

Simulation, define_functions, LinearChunked

Damysos.successful_retcode — Method

successful_retcode(x)

Returns true if x terminated with a successful return code.

Damysos.successful_retcode — Method

successful_retcode(path::String)

Loads .hdf5 file of convergence test and returns true if it was successful.

See also

ConvergenceTest

Damysos.terminated_retcode — Method

terminated_retcode(x)

Returns true if x terminated regularly.

Damysos.terminated_retcode — Method

terminated_retcode(path::String)

Loads .hdf5 file of convergence test and returns true if it terminated regularly.

See also

ConvergenceTest

Damysos.vx_cc — Method

vx_cc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vx|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_cc(h,1.0,-1.0)
0.5773502691896258

Damysos.vx_cv — Method

vx_cv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vx|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_cv(h,1.0,-1.0)
0.7886751345948129 - 0.21132486540518708im

Damysos.vx_vc — Method

vx_vc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vx|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_vc(h,1.0,-1.0)
0.7886751345948129 + 0.21132486540518708im

Damysos.vx_vv — Method

vx_vv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vx|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_vv(h,1.0,-1.0)
-0.5773502691896258

Damysos.vy_cc — Method

vy_cc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vy|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_cc(h,1.0,-1.0)
-0.5773502691896258

Damysos.vy_cv — Method

vy_cv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vy|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_cv(h,1.0,-1.0)
0.21132486540518708 - 0.7886751345948129im

Damysos.vy_vc — Method

vy_vc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vy|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_vc(h,1.0,-1.0)
0.21132486540518708 + 0.7886751345948129im

Damysos.vy_vv — Method

vy_vv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vy|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_vv(h,1.0,-1.0)
0.5773502691896258

Damysos.Δϵ — Method

Δϵ(h,kx,ky)

Returns the band energy (valence & conduction) difference at $\vec{k}=[k_x,k_y]$.

Damysos.ϵ — Method

ϵ(h,kx,ky)

Returns the eigenenergy of the positive (conduction band) state at $\vec{k}=[k_x,k_y]$.